MongoMapper mixins
